@@ -1,6 +1,6 @@
 {
   "name": "@gdjiami/app-template",
-  "version": "0.1.1",
+  "version": "0.2.1",
   "description": "template from jm-cli",
   "main": "index.js",
   "homepage": "https://github.com/carney520/jm-app-template#readme",
@@ -19,20 +19,17 @@
     "history": "^4.7.2",
     "mobx": "^5.1.0",
     "mobx-react": "^5.2.6",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-hot-loader": "4.3.12",
-    "react-loadable": "^5.5.0",
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2",
     "react-router": "^4.4.0-beta.6",
     "react-router-dom": "^4.4.0-beta.6",
-    "tslib": "1.9.3"
+    "tslib": "1.9.3",
+    "typescript": "^3.2.1"
   },
   "devDependencies": {
-    "@gdjiami/cli": "*",
-    "@types/react": "^16.7.6",
-    "@types/react-dom": "^16.0.9",
-    "@types/react-hot-loader": "^4.1.0",
-    "@types/react-loadable": "^5.4.1",
+    "@gdjiami/cli": "^0.1.24",
+    "@types/react": "^16.7.13",
+    "@types/react-dom": "^16.0.11",
     "@types/react-router-dom": "^4.3.1",
     "@types/webpack-env": "^1.13.6",
     "husky": "^1.1.4",
